virtual interview/requirements.txt
docx2txt==0.8
fitz==0.0.1.dev2
pandas==2.2.2
passlib==1.7.4
Pillow==10.3.0
plotly==5.18.0
pyttsx3==2.90
scikit_learn==1.4.0
spacy==3.7.2
SpeechRecognition==3.10.1
streamlit==1.31.0
